+++ MercadoPagoSDK/MercadoPagoSDK/UI/PXResult/PXPaymentCongratsViewModel.swift
@@ -329,18 +329,28 @@ extension PXPaymentCongratsViewModel: PXNewResultViewModelInterface {
         }
     }
 
+    func getDebinProperties() -> [String: Any]? {
+        return paymentCongrats.bankTransferProperties
+    }
+
     func getTrackingPath() -> PXResultTrackingEvents? {
         if let internalTrackingPath = paymentCongrats.internalTrackingPath as? PXResultTrackingEvents {
             return internalTrackingPath
         } else {
             var screenPath: PXResultTrackingEvents?
             let paymentStatus = paymentCongrats.type.getRawValue()
-            if paymentStatus == PXPaymentStatus.APPROVED.rawValue || paymentStatus == PXPaymentStatus.PENDING.rawValue {
-                screenPath = .congratsPaymentApproved(getTrackingProperties())
-            } else if paymentStatus == PXPaymentStatus.IN_PROCESS.rawValue {
-                screenPath = .congratsPaymentInProcess(getTrackingProperties())
+            var properties = getTrackingProperties()
+
+            if let debinProperties = getDebinProperties() {
+                properties.merge(debinProperties) { current, _ in current }
+            }
+
+            if paymentStatus == PXPaymentStatus.APPROVED.rawValue {
+                screenPath = .congratsPaymentApproved(properties)
+            } else if paymentStatus == PXPaymentStatus.IN_PROCESS.rawValue || paymentStatus == PXPaymentStatus.PENDING.rawValue {
+                screenPath = .congratsPaymentInProcess(properties)
             } else if paymentStatus == PXPaymentStatus.REJECTED.rawValue {
-                screenPath = .congratsPaymentRejected(getTrackingProperties())
+                screenPath = .congratsPaymentRejected(properties)
             }
 
             return screenPath