+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unpublished
 
 ### Added
+- Add bank_name and external_account_id for DEBIN tracks
 - New payment method model support
 - Added Congrats Extra Info (IFPE regulations)